Handling Unchanged TOAST Columns as a part MIRROR for CDC (#111)
**TOAST Storage:**

Postgres utilizes TOAST storage for large column values (>8KB) instead
of storing them directly in the table's heap (data pages). TOAST breaks
down the value into multiple pieces and stores them separately,
providing compression by default for optimization. The end-user is not
aware of this internal storage mechanism. Additional information about
TOAST can be found
[here](https://www.postgresql.org/docs/current/storage-toast.html).

**Problem:**

In Postgres, the logical decoding/replication feature is used by MIRROR
for CDC. Logical decoding provides the operation type (INSERT, DELETE,
UPDATE) and all column values for each row/tuple. However, if a column
value is a TOAST and remains unchanged during a data manipulation
language (DML) operation, Postgres does not provide the actual column
value. Instead, it offers a pointer that refers to the `pg_catalog.*`
table/chunk. It becomes the responsibility of the client to fetch the
value based on this pointer. This poses a challenge when normalizing raw
changes to the final (normalized) table, as TOAST columns can be
nullified.

**Solution:**

1. While reading the logical replication slot from Postgres, the
unchanged columns per tuple are set using the last seen column value in
that batch. This step is performed by **PullRecords**.
2. If the column value is not available in that batch, it is marked and
added to the unchanged toast column array, which is then propagated to
the raw table on the sink. Constructing the unchanged toast column array
is done in **PullRecords** and propagating the array to the raw table to
the sink is done in **SyncRecords**.
3. If the last seen column of a row is unchanged, the column value is
not updated. It remains the same as before the batch was normalized. To
achieve this, the UPDATE statements as part of the MERGE command in
**NormalizeRecords** are generated in a way that ensures unchanged
columns are not updated.

Step 1 can be a significant optimization for workloads that frequently
insert/update large values of the same row. This is particularly
applicable in IoT/NoSQL-like workloads. PeerDB controls the population
of unchanged column values per table based on the largest value seen in
that batch.