Nexus tests now idempotent (#229)
This PR aims to remove reliance on third party locations for our tests
and make them idempotent (i.e should run the same way every time and not
rely on any state across runs)

**Changes:**
- Peer configuration proto messages have json deserialisation traits
added to them.
- `cargo test` now first runs `CREATE PEER` commands for all the data
stores.
- For Postgres, we hydrate the test peer with the test table and its
data. The Postgres Peer used for testing is the same as catalog. the
test table is in a different schema - `test`. Creating it in a different
database involves connecting, creating the db and reconnecting. If
needed this can be done instead.
- ci.yaml creates the json files for the Bigquery and Snowflake peers in
an `assets` folder in `server/tests`.