added Rust and Go support for encrypted PKCS#8 private keys (#393)
Caveat: The Rust library we use doesn't support decrypted keys with
"weak" ciphers like DES out of the box, but is gated behind a feature
flag. The Snowflake guide for setting up private keys unfortunately
shows openssl commands that a weak cipher, so they will not work with
our Rust library.

A new configuration variable 'password' needs to be set when an
encrypted primary key is being used to set up a Snowflake peer, and
omitted otherwise.