PAUSE MIRROR support (#605)
`PAUSE MIRROR` supports both CDC and QRep mirrors. This is implemented
by signalling the Temporal workflow. A PauseSignal delivered to the
workflow should pause it, while a NoopSignal delivered to the workflow
should resume it. Signals that don't make sense for the current workflow
state are ignored. Currently supported only via the query layer, not
from the UI. The following caveats exist for now:

1) QRep mirrors cannot be paused during partition reading or
replicating, only after a run completes successfully and the signals are
read. This is usually not an issue but could be problematic for QRep
runs with a lot of partitions are running.
2) CDC mirrors similarly will only process signals and pause after a
pull + sync + normalize cycle is run. Due to 1), a CDC mirror cannot
pause during the initial snapshot phase, which is implemented as a QRep
mirror internally. This will be marked as unpausable in the UI.