Allow configurable column names for soft-delete and synced-at (#653)
In `soft_delete` mode, PeerDB creates a new `soft_delete_col_name`
column in each table in the destination. This mode is currently only
supported on Snowflake destination. Rows deleted that are deleted (based
on the primary key) are not deleted on the destination, rather the
column value is set to `true` in the `soft_delete` mode.