added spill to disk using Pebble for CDC records (#760)
This PR addresses cases where a large number of CDC records need to be
stored for record backfills, which can lead to either memory exhaustion
or having to provision large clusters.

After a certain point (currently 1M records stored, we only store the
latest record for each pkey) excess records are stored in a Pebble
database. Updates to keys already in-memory are continued to be stored
in-memory, all new keys are spilled to disk.