Migrate to Go 1.21's native slog (#764)
Uses `slog` and gets rid of `logrus`.
`flowName`, `deploymentUid` and for the most part `partitionId` are now
set at the top level function/workflow calls and are accessed by
children for logs. They do not need to be specified when logging. The
context values are read via logging middleware in `logger/handler.go`.