Improve loading: peers page (#840)
Convert /peers page to client components, so that we dont have to do a
round trip on every page visit. Converting it to client side rendering,
caches the page once it is loaded and we can load data via api. When
data is loading we show the loader. We also cache the data(via swr) and
refresh the data in background.