split BQ merge statements and run them individually (#1038)
For tables with a lot of columns in general and especially many toast
columns, some batches can trigger a MERGE statement so complex that
BigQuery is unable to process them, with errors like

`The query is too large. The maximum standard SQL query length is
1024.00K characters, including comments and white space characters.`
`Error 400: Resources exceeded during query execution: The query is too
complex., resourcesExceeded`

For now, the fix is splitting these complex MERGE statements into
smaller ones that act on different subsets of a raw table [partitioning
on the basis of `_peerdb_unchanged_toast_columns`]. This can lead to
tables need 10+ MERGE statements in a single batch, but this is a
compromise with our current design. Instead of sending MERGEs for all
tables at once, we do it per table now and update metadata at the end,
to avoid exceeding SQL query length limits.

---------

Co-authored-by: Kaushik Iska <iska.kaushik@gmail.com>