Support more data types (#1089)
This PR adds support for the following types across PG, SF and BQ:
- Array of Boolean
- Array of Date
- Array of Timestamp(/TZ)
- Array of Int16 (smallint)
- CIDR
- MacAddr
- INET

The latter 3 is more relevant for Postgres.
This PR achieves the correct AVRO type mapping for Date in BigQuery -
eliminating the hack which was being done in merge transforms.
This PR also writes date and time values as human readable text for BQ
QRep

Tests added as well

Fixes #666 
Fixes #20