dynamically add new tables to CDC mirrors (#1084)
1. Call an activity to add the new tables to the publication.
2. Kick off a child `CDCFlow` with only the new tables as input along
with the `InitialCopyOnly` flag set. This runs enough of the `SetupFlow`
and `SnapshotFlow` to be fine, but also skips over some checks regarding
primary keys and replica identities. The current idea is to have them in
the UI instead, just like in the create mirror interface.
3. Patch the config of the parent workflow with the new tables and then
resume.

Currently there is no way to interface with this feature, will add a way
to signal this in a different PR. UI interface will be added
subsequently.

---------

Co-authored-by: Kaushik Iska <iska.kaushik@gmail.com>