BigQuery cross project support: unqualify tables for cdc/qrep (#1147)
`DefaultDatasetID` and `DefaultProjectID` work only for unqualified
tables. This PR unqualifies our tables in bigquery queries for qrep and
cdc.

Will update our docs to explicitly say that we support cross-project
mode only for CDC and QRep (BQ). This is yet to be integrated for resync
mirror

This PR has been functionally tested with two projects.

Follow up to #1073