UI for editing mirror parameters (#1190)
`IdleTimeoutSeconds` and `SyncBatchSize` can be changed anytime, tables
can only be added IF mirror is paused, otherwise the button will be
disabled.

Additional changes for this:
1) moved CDC dynamic properties to the main signal selector and blocking
on the main signal selector while paused, so we can receive mirror edits
even while paused.
2) pass `TableNameSchemaMapping` from state always to `NormalizeFlow` in
case new tables were added.
3) added a new property to the tables component to disable selecting of
tables already in the mirror.

There is a bug where tables added to the mirror dynamically will not
reflect in the status screen, will be fixed in a follow up PR.
`TableMappings` Needs to be stored in state as well.