explicitly set client_encoding to UTF8 (#1236)
A couple of our queries [`getTableSchemaForTable` and
`getChildToParentRelIDMap`] use the simple query protocol (why is this
needed? not sure). pgx requires `client_encoding` to be set to `UTF8`
for using the simple query protocol, but it won't be set if the database
we are connecting to doesn't use the `UTF8` encoding, which isn't very
likely but still possible.

Fixed by explicitly mentioning `client_encoding` in the connection
string. Also moving `application_name` there.