BigQuery: Skip merge if no primary keys present (#1290)
Merge statements currently are syntactically incorrect for tables
without a primary key. This case is ideally not seen as tables without
pkey/replica identity full are filtered out the UI level but in branches
where we rely on signals for adding tables, this case can go unchecked
and normalize flow fails.

This PR skips a table during merge if there are no primary keys for it.