Clickhouse: remove nullable mapping (#1344)
It appears timeseries performance is a common use-case for Clickhouse,
wherein users often index based on a `DateTime` column. Clickhouse does
not allow indexing based on `Nullable` columns. Also having nullable
columns decreases performance.
Thus, this PR chooses to stay with the original mapping of DateTime and
Date (without the Nullable)

A decision is to be made in the future regarding whether we get user's
preference on setting a column as Nullable or not for Clickhouse mirrors