Validate peer: fix snowflake validate (#1447)
Snowflake validate could fail if the user creates an internal schema
before hand, in which case permissions for schema creation are not
necessary and may not be given. Our validate for snowflake will fail as
it tries to create a dummy schema

The change this PR makes is to check if the raw schema exists (see
note). If it doesn't then, create it. The remaining steps - table
create, insert row etc. remain the same. `create schema if not exists`
does not bypass snowflake's permission checks for schema creation

Removes drop schema step as that isn't something we do in mirrors

note - this is also done at raw table creation and qrep metadata table
creation where we create schemas