comparing RelationMessage against TableSchema, stop caching RelationMessages (#1438)
Previously we diffed the previous and current `RelationMessages`
received from Postgres during logical replication to ascertain if the
schema changed or not. This fails in cases where the schema changed
after mirror setup but before we get the first `RelationMessage` which
can happen a) during initial load, before CDC and b) during CDC, if no
rows are synced before schema changes are made.

This causes a desync between PeerDB's view of the table and the actual
status of the table, causing issues. Changed to diff from the current
`TableSchema` instead, which should always be present. Eliminates having
to store and pass `RelationMessageMapping` around, it is scoped to
reading from the replication connection.

Also fixes bugs with schema changes being handled incorrectly for PG and
BQ, and schema audits being logged without schema deltas actually being
present.