column exclusion - properly handle schema changes (#1512)
Column exclusions works by removing columns from the schema we fetch of
the source table. This exclusion was not being done in the code path for
schema changes [where we fetch the schema again], causing a disconnect
and normalize to fail.

Fixed by moving the exclusion code to a separate function and making
both code paths use it. Also CDC handles excluded columns earlier to
prevent spurious logs.