QRep: Adjust start value for TID watermark column (#1569)
This issue was noticed with tables where CTIDs are _not_ unique ex:
partitioned tables
We adjust(increment) the start values of partitions after the first
partition. This is to avoid overlapping partitions.

This was not done for TID columns because the assumption was that they
would always be unique.
As a result, this could lead to duplication of rows for initial load and
XMIN where CTIDs are the watermark column.