BigQuery Add Column: Account for update limit (#1662)
BigQuery has a
[quota](https://cloud.google.com/bigquery/quotas#standard_tables) of 5
metadata update operations per table per 10 seconds. So currently if
many columns are added in a PG to BQ CDC mirror, sync records fails with
this error from BigQuery - even though we do `ADD COLUMN IF NOT EXISTS`
(it still counts as an update operation)

This PR accounts for this by manually checking the table metadata for
the column before adding it. This will atleast ensure that on sync flow
retry, the remaining columns get added

Functionally tested