feat(cdc): reload peers on flow resume (#1763)
Implement the reloadPeers function to support the edit peer
functionality. When a flow is paused and resumed, reload the latest peer
details for both the source and destination peers. This ensures that any
changes made to the peer configuration during the pause are picked up
when the flow resumes execution.