Add query tags for Snowflake connector to track most expensive queries
This commit adds query tags to the Snowflake connector for the most frequently executed and expensive queries that contribute significantly to the overall cost. The tags are added using the `withMirrorNameQueryTag` method, which sets a tag with the format `peerdb-mirror-<flowJobName>` for each query.

Query tags have been added to key methods:
- `SyncRecords`
- `NormalizeRecords`
- `CreateRawTable`
- `SyncFlowCleanup`
- `SyncQRepRecords`
- `SetupQRepMetadataTables`
- `ConsolidateQRepPartitions`

These methods cover core functionality of syncing, normalizing, and managing QRep tasks. The tags will help identify and track the most expensive queries in the Snowflake web interface or using the `QUERY_HISTORY` table.

Some minor queries have been ignored to focus on the most impactful areas. The query tags will aid in monitoring, optimizing, and controlling costs associated with the Snowflake connector.