BigQuery rename tables: account for json column in replica identity full table (#1795)
When handling soft deletes, we perform equality checks for primary key
columns in _resync and original tables. For a replica identity full
table, we mark all columns as primary keys.
If one of the columns in that case is a JSON, `=` comparison fails. Need
to use TO_JSON_STRING instead like how we do for merge.

Functionally tested