UI: fix dst table overwrite in QRep mirror creation (#1799)
When the watermark/source table changed, we had logic to reset the
destination table name to the watermark table name. This was done
improperly, causing a desync between the displayed value to the user and
the actual value in config that the mirror is created with.

Fixed by removing the logic entirely, users should manually change the
table if watermark table is updated after destination table.