Drop mirror : Drop stats option (#2029)
![Screenshot 2024-09-02 at 6 36
57 PM](https://github.com/user-attachments/assets/340c30c6-47f8-42a5-990d-9624407385fc)

This PR:
- Adds drop mirror stats (present in `peerdb_stats` schema)
functionality to drop mirror. The failure of this step is not fatal to
`drop_flow`. Exposed in UI as seen above.
- Resync now drops mirror stats
- In query layer, drop mirror does not drop stats. Extending the drop
mirror command syntax to support this will be done if there's demand

Also:
- Adds a column - `parent_mirror_name (TEXT)` to `qrep_runs` and
`qrep_partitions` tables.
- Adds a field - `parentMirrorName (string)` to QRep Config proto
- Adaps `monitoring.go` and initial load stats endpoint to use parent
mirror name

Parent mirror name is the name of the parent peer flow in case of
initial load clone. It is the same as flow name for QRep and XMin
mirrors

- [ ] Functionally tested