Increase retry initial backoff to 1 minute (#2418)
Retrying in 1 sec very rarely changes things as we typically tend to
wait for database auto-scaling, network issue resolution or other error
types wherre we would anyways need to wait in the order of minuted.

It was also causing alert spamming in the systems basically making them
not usable for a certain kind of errors.