ClickHouse: Support for special characters in column names (#2705)
Avro requires alphanumeric + underscore field names. However Postgres
column names do not have the same restriction.

- This PR introduces a map of column names to an Avro compatible
encoding of the same. Violating characters are replaced by underscore
and the resulting string is suffixed with a serial number.
- This is utilized only for the initial load Avro loading since for CDC
the columns will be that of our raw table columns.
- This PR only implements this support for ClickHouse target peers
- It also supports these columns for our normalize inserts by escaping
the single quotes in column name argument of the JSONExtract* functions
- Functionally tested and the unprivileged columns E2E test has been
adapted to include more columns to test this PR

Fixes #2701