ClickHouse Initial Load: Error out if we see a column on ClickHouse which is not on source (#2793)
In a case where a customer syncs to an existing destination table which
has a column not on the source table, initial load currently fails
because we get the columns from ClickHouse to design the Insert into
Select from Avro - which has source columns as fields.
However the error currently surfaces as a vague syntax error where there
is an empty string column name in the command.

This PR just catches this case and errors out more explicitly

A follow-up to this is passing excluded columns info to this function so
that we don't have to get the schema from ClickHouse

---------

Co-authored-by: Kevin Biju <52661649+heavycrystal@users.noreply.github.com>