feat(classifier): add ACCESS_ENTITY_NOT_FOUND with condition as recoverable (#2833)
When droppping a pipe during initial load it can take some time for the
workflow to get cancelled. During this time the insert queries fail with
errros like: `ID(a14c2a1c-edcd-5fcb-73be-bd04e09fccb7) not found in user
directories`.

For now classifying as recoverable because if seen for a longer duration
we should be alerted


ref https://github.com/ClickHouse/ch-go/pull/1058 for adding the new
error code