maria ci: fix heisenbug (#2855)
https://dev.mysql.com/doc/refman/8.4/en/set-variable.html

> If you change a global system variable, the value is remembered and
used to initialize the session value for new sessions until you change
the variable to a different value or the server exits. The change is
visible to any client that accesses the global value. However, the
change affects the corresponding session value only for clients that
connect after the change. **The global variable change does not affect
the session value for any current client sessions (not even the session
within which the global value change occurs).**