feat(wf-status): raise error if flow not found in catalog (#2939)
During maintenance if snapshot flow is dropped, it can sometimes not
update the temporal status.
This leads to missing flow entry in catalog once drop succeeds and
GetWorkflowStatus falls back to temporal which shows stale status
(snapshot).
If instead we get `pgx.ErrNoRows` we can raise the error further and
assume there is a bigger issue at hand
- in this case, maintenance would retry the activity and next fetch
should not have the mirror in the list at all