Logs tab in UI: fix extractFromCloneName (#2926)
In order to make it more apparent that a log from a "clone_" mirror is
actually an initial load of its parent, in the UI logs tab we do regex
parsing to get the parent name and say (initial load).
However the regex parsing doesn't work if the parent name has
underscores. This PR fixes this via a stopgap regex fix.
Long term would want to introduce and populate a parent name column in
the logs table similar to the qrep stats tables