move type conversion related qvalue package to shared (#3011)
Breaking out the task to more manageable steps to reduce risk of
introducing regression.

For now I am focusing on the logic needed to get supported type
conversions only (and not worrying about displaying the default type
mapping, which is a lot hairier).

Approach is to focus on migrating `ListSupportedTypeConversions` (and
packages it depends on) to shared, and refactor code so that the
packages migrated to shared/ does not import packages from the rest of
peerdb codebase.

Test: CI + locally tested a simple flow.