ClickHouse: Maintain normalize metadata to prevent redundant sync steps (#2662)
Consider a mirror with 5 tables. Currently, if a push to one of those 5
tables fails, we will retry normalize, and this retry, the following
redundant steps occur:
1) **We load the S3 files to the raw table again**: This can be
expensive (time, memory, storage, cost etc.) in cases where syncs have
gone way ahead of normalize.
2) **The remaining 4 tables will have these batches synced again**: The
inconveniences caused by this in terms of resources and data modelling
is beyond the scope of this PR description.

![Screenshot 2025-03-05 at 2 06
46 PM](https://github.com/user-attachments/assets/d611adc9-81a6-4e1a-8039-5e7f058babdf)


This PR introduces two new columns in the `metadata_last_sync_state`
column:
1) A bigint column which denotes the most recently synced batch id in
raw table
2) A json column which tells us the recent batch id normalized to each
target table of the mirror.

We use this to memoize our pushes to raw table and the final tables

![Screenshot 2025-03-05 at 2 19
38 PM](https://github.com/user-attachments/assets/d6a7d276-41d4-4f99-9fe3-67fff19ecfbb)


#### Notes
1) This can perhaps be implemented for other target peers too as a
follow up
3) This PR has been functionally tested and seen to work. **This will be
landed after a scale+chaos test**

---------

Co-authored-by: Philip Dubé <philip.dube@clickhouse.com>