change QValueTime/QValueTimeTZ to internally use time.Duration not time.Time (#3036)
MySQL should instead map `time` to `QValueTimestamp`,
since `QValueTime` maps to types on destinations where time is expected to be time part of datetime
(ie, from 00:00 up to 24:00)

Keep existing QValueTime handling to not break existing mirrors.
Note existing mirrors have a problem where in cdc they will have time be modulos 24 hours

Another issue was that postgres time values over cdc on clickhouse were coming in with current year instead of unix epoch. This is because we relied on `parseDateTime64BestEffort`,
which interprets times without dates as happening in current year (or previous year in some cases).
Adjust normalization to have year be 1970 in order to be consistent between qrep & cdc

This also helps allow in the future mapping QValueTime to DateTime instead of DateTime64

Adjust tests to not be so permissive, which let these discrepancies exist, but which going forward we should prevent