mongo: handle large mongo document (#3108)
This PR handles two issues w.r.t. large event (separated by commits,
recommend review them separately):

1) Mongo limits document size to 16MB, trying to insert even one byte
larger will result in error:
>  [object to insert too large. size in bytes: xxx, max size: xxx]

With this information I was able to generate document in test that is
exactly the maximum size that we can write to mongodb -- and we should
support up to this size in our CDC connector without errors.

Currently we seem to be imposing snowflake's 15MB limit in the shared
`MarshalJSONWithOptions` function, so the e2e test added to this PR was
failing because the json got set to "{}". With this PR, Mongo is made an
exception (although i think the current implementation is sloppy since
it's based on source data type rather than destination datatypes).
Looking for feedback on a better solution, or if we can get ride of the
15MB limit altogether in `MarshalJSONWithOptions` given `QValueToAvro`
seem to handle snowflake's edge cases already -- but not sure if I'm
missing something)

2) With above change, inserts/replace/delete events processed
successfully during CDC, but update events was still failing because it
was double the size as the other events and change stream also has a
16MB limitation on event size (because the value was present at least
twice in the event: once in the `fullDocument` field and once in the
`updateDescription` field, not to mention it could also be present in
the `fullDocumentBeforeChange` field if available). So the PR:
- set `fullDocumentBeforeChange` to `off` always, which we are not using
- make the pipeline filter out `updateDescription` field, which we are
not using
- restrict to only event type we care about so far
(insert/update/delete/replace) and only to fields we need
(operatinType/clusterTime/documentKey/fullDocument/ns) so far.

Intentionally not using `$changeStreamSplitLargeEvent` for now as it
seems quite discouraged by [mongo
doc](https://www.mongodb.com/docs/manual/reference/operator/aggregation/changeStreamSplitLargeEvent/)
-- we can support it if current implementation is no longer sufficient.

Test: e2e test passing with this change.